+++ plugins/kubernetes/dev/index.tsx
@@ -33,6 +33,7 @@ import fixture1 from '../src/__fixtures__/1-deployments.json';
 import fixture2 from '../src/__fixtures__/2-deployments.json';
 import fixture3 from '../src/__fixtures__/1-cronjobs.json';
 import fixture4 from '../src/__fixtures__/2-cronjobs.json';
+import fixture5 from '../src/__fixtures__/1-rollouts.json';
 import { TestApiProvider } from '@backstage/test-utils';
 
 const mockEntity: Entity = {
@@ -179,6 +180,19 @@ createDevApp()
     element: (
       <TestApiProvider
         apis={[[kubernetesApiRef, new MockKubernetesClient(fixture4)]]}
+      >
+        <EntityProvider entity={mockEntity}>
+          <EntityKubernetesContent />
+        </EntityProvider>
+      </TestApiProvider>
+    ),
+  })
+  .addPage({
+    path: '/fixture-5',
+    title: 'Fixture 5',
+    element: (
+      <TestApiProvider
+        apis={[[kubernetesApiRef, new MockKubernetesClient(fixture5)]]}
       >
         <EntityProvider entity={mockEntity}>
           <EntityKubernetesContent />