+++ plugins/notifications/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @backstage/plugin-notifications
 
+## 0.2.4-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.14.10-next.0
+  - @backstage/core-plugin-api@1.9.3
+  - @backstage/errors@1.2.4
+  - @backstage/theme@0.5.6
+  - @backstage/types@1.1.1
+  - @backstage/plugin-notifications-common@0.0.5
+  - @backstage/plugin-signals-react@0.0.4
+
+## 0.2.3
+
+### Patch Changes
+
+- 3bf0697: The toolbar on the Notifications page is hidden when there are no listed notifications.
+- Updated dependencies
+  - @backstage/core-components@0.14.9
+  - @backstage/plugin-notifications-common@0.0.5
+  - @backstage/core-plugin-api@1.9.3
+  - @backstage/errors@1.2.4
+  - @backstage/theme@0.5.6
+  - @backstage/types@1.1.1
+  - @backstage/plugin-signals-react@0.0.4
+
+## 0.2.3-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.14.9-next.1
+
+## 0.2.3-next.1
+
+### Patch Changes
+
+- 3bf0697: The toolbar on the Notifications page is hidden when there are no listed notifications.
+- Updated dependencies
+  - @backstage/core-components@0.14.9-next.0
+  - @backstage/core-plugin-api@1.9.3
+  - @backstage/errors@1.2.4
+  - @backstage/theme@0.5.6
+  - @backstage/types@1.1.1
+  - @backstage/plugin-notifications-common@0.0.4
+  - @backstage/plugin-signals-react@0.0.4
+
 ## 0.2.3-next.0
 
 ### Patch Changes